builder pattern all the way through
